done 4.20-4.23

+++ part4/blogList/controllers/blogs.js
@@ -17,17 +17,13 @@ blogRouter.get('/', async (req, res, next) => {
 
 blogRouter.post('/', async (req, res, next) => {
   const body = req.body
-
+  
   try {
-    const decodedToken = jwt.verify(req.token, process.env.SECRET)
-    if (!decodedToken.id) {
-      return res.status(401).json({ error: 'token invalid' })
+    const user = req.user // get specific user who holding the token from userExtractor middleware
+    if (!user) {
+      return res.status(401).json({ error: 'token invalid or user not found'})
     }
 
-    const user = await User.findById(decodedToken.id)
-    if (!user) {
-      return res.status(400).json({ error: 'user not found' })
-    }
 
     const blog = new Blog({
       title: body.title,
@@ -48,8 +44,22 @@ blogRouter.post('/', async (req, res, next) => {
 })
 
 blogRouter.delete('/:id', async (req, res, next) => {
-  const id = req.params.id
+  const id = req.params.id  
   try {
+    const user = req.user
+    if (!user) {
+      return res.status(401).json({ error: 'token invalid or user not found'})
+    }
+    
+    const blog = await Blog.findById(id)
+    if (!blog) {
+      return res.status(404).json({ error: 'blog not found' })
+    }
+    
+    if (blog.user.toString() !== user.id.toString()) {
+      return res.status(403).json({ error: 'permission denied' })
+    }
+
     await Blog.findByIdAndDelete(id)
     res.status(204).end()
   } catch (err) {
